adjusting puzzle_solver to listen to camera_node and send feedback to robot, removed hand made camera_node

+++ src/puzzle_solver/include/puzzle_processing.h
@@ -19,7 +19,7 @@ using namespace cv;
 using std::vector;
 using std::pair;
 
-extern int PUZZLE_SIZE;
+extern size_t PUZZLE_SIZE;
 extern Size PUZZLE_IMAGES_SIZE;
 extern vector<Mat> initialPuzzleImages;
